добавлена библиотека h5py в рекомендации
