+++ screenqual/models/banner_classification/neural_network.py
