+++ tests/regression/banner_detector/__init__.py
