+++ tests/regression/banner_detector/test_regression_banner_detector.py
@@ -4,5 +4,5 @@ from tests.regression.regression_test import TestRegression
 
 class TestRegressionBanner(TestRegression):
     def test_fscore_banner_images(self):
-        self.fscore(["/banner/with_banner/"], ["/banner/without_banner/"],
-                    BannerAnalyser(), .9, "banner detector", extension="jpg")
+        self.fscore(["/banner_detector/with_banner/"], ["/banner_detector/without_banner/"],
+                    BannerAnalyser(), .0, "banner_detector detector", extension="jpg")
