+++ tests/unit/banner_detector/img/with_banner/1.jpg
