+++ tests/unit/banner_detector/img/with_banner/2.jpg
