+++ tests/unit/banner_detector/img/with_banner/3.jpg
