+++ tests/unit/banner_detector/img/with_banner/4.jpg
