+++ tests/unit/banner_detector/img/without_banner/1.jpg
