+++ tests/unit/banner_detector/img/without_banner/3.jpg
